+++ src/Component/Collector.php
@@ -48,10 +48,6 @@ class Collector
                     $collection->setMiddleware($component->get());
 
                     break;
-                default:
-                    self::set($component, $container);
-
-                    break;
             }
 
             $component->set($container);
@@ -64,27 +60,5 @@ class Collector
         }
 
         return $collection;
-    }
-
-    /**
-     * Sets the component and add it to the container.
-     * 
-     * @param  \Rougin\Slytherin\Component\ComponentInterface $component
-     * @param  \Interop\Container\ContainerInterface &$container
-     * @return void
-     */
-    private static function set(ComponentInterface $component, ContainerInterface & $container)
-    {
-        $componentContainer = $component->needsContainer() ? $container : null;
-        $componentName = (string) $component;
-        $component = $component->get($componentContainer);
-
-        if (is_array($component)) {
-            foreach ($component as $key => $value) {
-                $container->add($key, $value);
-            }
-        } else {
-            $container->add($componentName, $component);
-        }
     }
 }