+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to `Slytherin` will be documented in this file
 
+## [Unreleased](https://github.com/rougin/slytherin/compare/v0.6.0...HEAD)
+
+### Fixed
+- Using `add` in `Component\Collector` if not using `IoC\Vanilla\Container`
+
 ## [0.7.0](https://github.com/rougin/slytherin/compare/v0.6.0...v0.7.0)
 
 ### Added